Limit images in content to 95vh.
It may occur (especially on wide screens) that some image is higher
than the viewport making it impossible to see the image as whole while
scrolling past. (It could be done by clicking on it and viewing it
in the gallery.)

I personally think that whole image outweighs maximal
width. So, limit the image height with `max-width` rule to `95vh`
(95% viewport's heigh). Also replace the hard-coded `width` of `90%`
width a `max-width` so that the image can scale down to satisfy the
`max-height`. Otherwise it stretches and uses the default object-fit of
`cover`  defeating the whole purpose of this commit.